Add ssh-auth-sock to docs example
I found out that to make this action to work properly I had to set `ssh-auth-sock: ${{ github.workspace }}/ssh-auth.sock` in the ssh action config.

I'm not 100% sure why it's needed but thought it might be good to mention it in the documentation more properly. Maybe you know more why it's needed. And also, feel free to skip it if you think this has more to do with my setup than with the actual project!